+++ docker/containerlist.go
@@ -25,5 +25,8 @@ func (d *Docker) GetContanerList() ([]string, error) {
 		}
 	}
 
+	/*if len(containerList) == 0 {
+		return nil, errors.New("No running projects. Start project using docker-compose up and then run again `jumper start`")
+	}*/
 	return containerList, nil
 }